Refactor Windows settings UI
Split the monolithic Settings window into lazy-loaded section controls and defer analytics and media device initialization until needed.
