+++ windows/src/TinyClips.App/Settings/Sections/AboutSettingsSection.xaml
@@ -1,0 +1,96 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<UserControl
+    x:Class="TinyClips.App.Settings.Sections.AboutSettingsSection"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d">
+
+    <StackPanel Spacing="8">
+        <TextBlock
+            Margin="4,8,0,4"
+            Style="{StaticResource SubtitleTextBlockStyle}"
+            Text="About" />
+
+        <Border Style="{StaticResource SettingCardStyle}">
+            <StackPanel Spacing="4">
+                <TextBlock Style="{StaticResource BodyStrongTextBlockStyle}" Text="Tiny Clips for Windows" />
+                <TextBlock
+                    x:Name="AboutVersionText"
+                    Style="{StaticResource CaptionTextBlockStyle}"
+                    Foreground="{ThemeResource TextFillColorSecondaryBrush}"
+                    Text="Version" />
+                <TextBlock
+                    Style="{StaticResource CaptionTextBlockStyle}"
+                    Foreground="{ThemeResource TextFillColorSecondaryBrush}"
+                    Text="A fast, beautiful screen capture app for screenshots, video, and GIFs." />
+            </StackPanel>
+        </Border>
+
+        <Border
+            x:Name="DirectBuildUpdatesCard"
+            Visibility="Collapsed"
+            Style="{StaticResource SettingCardStyle}">
+            <StackPanel Spacing="4">
+                <TextBlock Style="{StaticResource BodyStrongTextBlockStyle}" Text="Updates (Direct build)" />
+                <TextBlock
+                    Style="{StaticResource CaptionTextBlockStyle}"
+                    Foreground="{ThemeResource TextFillColorSecondaryBrush}"
+                    Text="This build updates through winget (`winget upgrade Refractored.TinyClips`). Tiny Clips does not include an in-app updater." />
+            </StackPanel>
+        </Border>
+
+        <Border
+            x:Name="StoreBuildUpdatesCard"
+            Visibility="Collapsed"
+            Style="{StaticResource SettingCardStyle}">
+            <StackPanel Spacing="4">
+                <TextBlock Style="{StaticResource BodyStrongTextBlockStyle}" Text="Updates (Microsoft Store build)" />
+                <TextBlock
+                    Style="{StaticResource CaptionTextBlockStyle}"
+                    Foreground="{ThemeResource TextFillColorSecondaryBrush}"
+                    Text="This build is updated by Microsoft Store auto-update; direct/winget update UI is intentionally hidden." />
+            </StackPanel>
+        </Border>
+
+        <Border Style="{StaticResource SettingCardStyle}">
+            <Grid ColumnSpacing="16">
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="*" />
+                    <ColumnDefinition Width="Auto" />
+                </Grid.ColumnDefinitions>
+                <StackPanel VerticalAlignment="Center">
+                    <TextBlock Style="{StaticResource BodyStrongTextBlockStyle}" Text="Source code" />
+                    <TextBlock
+                        Style="{StaticResource CaptionTextBlockStyle}"
+                        Foreground="{ThemeResource TextFillColorSecondaryBrush}"
+                        Text="View the project, open an issue or feature request, and contribute on GitHub." />
+                </StackPanel>
+                <StackPanel
+                    Grid.Column="1"
+                    VerticalAlignment="Center"
+                    HorizontalAlignment="Right"
+                    Spacing="4">
+                    <HyperlinkButton
+                        AutomationProperties.AutomationId="AboutRepoLink"
+                        AutomationProperties.Name="Open the Tiny Clips GitHub repository"
+                        Content="GitHub repository"
+                        NavigateUri="https://github.com/jamesmontemagno/tiny-clips" />
+                    <HyperlinkButton
+                        x:Name="AboutIssueLink"
+                        AutomationProperties.AutomationId="AboutIssueLink"
+                        AutomationProperties.Name="Open an issue or feature request on GitHub"
+                        Content="Open an issue or feature request" />
+                </StackPanel>
+            </Grid>
+        </Border>
+
+        <TextBlock
+            x:Name="AboutCopyrightText"
+            Margin="4,4,0,0"
+            Style="{StaticResource CaptionTextBlockStyle}"
+            Foreground="{ThemeResource TextFillColorSecondaryBrush}"
+            Text="© Refractored LLC" />
+    </StackPanel>
+</UserControl>
