+++ windows/src/TinyClips.App/Settings/Sections/GeneralSettingsSection.xaml.cs
@@ -1,0 +1,34 @@
+using System;
+using Microsoft.UI.Xaml;
+using Microsoft.UI.Xaml.Controls;
+
+namespace TinyClips.App.Settings.Sections;
+
+/// <summary>
+/// General settings: theme, save location, file naming, launch-at-login, and capture behavior
+/// toggles.
+/// </summary>
+public sealed partial class GeneralSettingsSection : UserControl
+{
+    private readonly IDisposable _realizationScope;
+
+    public SettingsViewModel ViewModel { get; }
+
+    /// <summary>
+    /// Raised when the user clicks Browse. The folder picker must be owned by the shell window
+    /// (it needs an HWND via <c>WinRT.Interop.WindowNative</c>), so this section only requests it
+    /// rather than showing a picker itself.
+    /// </summary>
+    public event EventHandler? BrowseSaveDirectoryRequested;
+
+    public GeneralSettingsSection(SettingsViewModel viewModel)
+    {
+        ViewModel = viewModel;
+        _realizationScope = viewModel.BeginSectionRealization();
+        InitializeComponent();
+        SectionLifecycle.HookFirstLoad(this, viewModel, _realizationScope);
+    }
+
+    private void OnBrowseSaveDirectory(object sender, RoutedEventArgs e) =>
+        BrowseSaveDirectoryRequested?.Invoke(this, EventArgs.Empty);
+}
