+++ windows/src/TinyClips.App/Settings/Sections/GifSettingsSection.xaml
@@ -1,0 +1,149 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<UserControl
+    x:Class="TinyClips.App.Settings.Sections.GifSettingsSection"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d">
+
+    <StackPanel Spacing="8">
+        <TextBlock
+            Margin="4,8,0,4"
+            Style="{StaticResource SubtitleTextBlockStyle}"
+            Text="GIF" />
+
+        <TextBlock
+            Margin="4,4,0,0"
+            Style="{StaticResource BodyStrongTextBlockStyle}"
+            Text="GIF quality" />
+
+        <Border Style="{StaticResource SettingCardStyle}">
+            <Grid ColumnSpacing="16">
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="*" />
+                    <ColumnDefinition Width="Auto" />
+                </Grid.ColumnDefinitions>
+                <StackPanel VerticalAlignment="Center">
+                    <TextBlock Style="{StaticResource BodyStrongTextBlockStyle}" Text="Frame rate" />
+                </StackPanel>
+                <NumberBox
+                    Grid.Column="1"
+                    Width="120"
+                    VerticalAlignment="Center"
+                    AutomationProperties.AutomationId="GifFrameRateNumberBox"
+                    AutomationProperties.Name="GIF frame rate"
+                    Maximum="50"
+                    Minimum="5"
+                    SpinButtonPlacementMode="Inline"
+                    Value="{x:Bind ViewModel.GifFrameRate, Mode=TwoWay}" />
+            </Grid>
+        </Border>
+
+        <Border Style="{StaticResource SettingCardStyle}">
+            <Grid ColumnSpacing="16">
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="*" />
+                    <ColumnDefinition Width="Auto" />
+                </Grid.ColumnDefinitions>
+                <StackPanel VerticalAlignment="Center">
+                    <TextBlock Style="{StaticResource BodyStrongTextBlockStyle}" Text="Maximum width" />
+                    <TextBlock
+                        Style="{StaticResource CaptionTextBlockStyle}"
+                        Foreground="{ThemeResource TextFillColorSecondaryBrush}"
+                        Text="GIF frames are scaled down to at most this width (pixels)." />
+                </StackPanel>
+                <NumberBox
+                    Grid.Column="1"
+                    Width="120"
+                    VerticalAlignment="Center"
+                    AutomationProperties.AutomationId="GifMaxWidthNumberBox"
+                    AutomationProperties.Name="GIF maximum width"
+                    Maximum="1920"
+                    Minimum="240"
+                    SpinButtonPlacementMode="Inline"
+                    Value="{x:Bind ViewModel.GifMaxWidth, Mode=TwoWay}" />
+            </Grid>
+        </Border>
+
+        <TextBlock
+            Margin="4,16,0,0"
+            Style="{StaticResource BodyStrongTextBlockStyle}"
+            Text="Recording &amp; output" />
+
+        <Border Style="{StaticResource SettingCardStyle}">
+            <Grid ColumnSpacing="16">
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="*" />
+                    <ColumnDefinition Width="Auto" />
+                </Grid.ColumnDefinitions>
+                <StackPanel VerticalAlignment="Center">
+                    <TextBlock Style="{StaticResource BodyStrongTextBlockStyle}" Text="Countdown before recording" />
+                </StackPanel>
+                <StackPanel
+                    Grid.Column="1"
+                    VerticalAlignment="Center"
+                    Orientation="Horizontal"
+                    Spacing="12">
+                    <NumberBox
+                        Width="120"
+                        AutomationProperties.AutomationId="GifCountdownDurationNumberBox"
+                        AutomationProperties.Name="GIF countdown seconds"
+                        IsEnabled="{x:Bind ViewModel.GifCountdownEnabled, Mode=OneWay}"
+                        Maximum="10"
+                        Minimum="1"
+                        SpinButtonPlacementMode="Inline"
+                        Value="{x:Bind ViewModel.GifCountdownDuration, Mode=TwoWay}" />
+                    <ToggleSwitch
+                        AutomationProperties.AutomationId="GifCountdownToggle"
+                        AutomationProperties.Name="Enable GIF countdown"
+                        IsOn="{x:Bind ViewModel.GifCountdownEnabled, Mode=TwoWay}" />
+                </StackPanel>
+            </Grid>
+        </Border>
+
+        <Border Style="{StaticResource SettingCardStyle}">
+            <Grid ColumnSpacing="16">
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="*" />
+                    <ColumnDefinition Width="Auto" />
+                </Grid.ColumnDefinitions>
+                <StackPanel VerticalAlignment="Center">
+                    <TextBlock Style="{StaticResource BodyStrongTextBlockStyle}" Text="Open trimmer after recording" />
+                    <TextBlock
+                        Style="{StaticResource CaptionTextBlockStyle}"
+                        Foreground="{ThemeResource TextFillColorSecondaryBrush}"
+                        Text="Drop frames from the start and end before saving." />
+                </StackPanel>
+                <ToggleSwitch
+                    Grid.Column="1"
+                    VerticalAlignment="Center"
+                    AutomationProperties.AutomationId="ShowGifTrimmerToggle"
+                    AutomationProperties.Name="Open GIF trimmer after recording"
+                    IsOn="{x:Bind ViewModel.ShowGifTrimmer, Mode=TwoWay}" />
+            </Grid>
+        </Border>
+
+        <Border Style="{StaticResource SettingCardStyle}">
+            <Grid ColumnSpacing="16">
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="*" />
+                    <ColumnDefinition Width="Auto" />
+                </Grid.ColumnDefinitions>
+                <StackPanel VerticalAlignment="Center">
+                    <TextBlock Style="{StaticResource BodyStrongTextBlockStyle}" Text="Copy GIF to the clipboard" />
+                    <TextBlock
+                        Style="{StaticResource CaptionTextBlockStyle}"
+                        Foreground="{ThemeResource TextFillColorSecondaryBrush}"
+                        Text="Copy the recorded GIF file so it can be pasted into chats and editors." />
+                </StackPanel>
+                <ToggleSwitch
+                    Grid.Column="1"
+                    VerticalAlignment="Center"
+                    AutomationProperties.AutomationId="CopyGifToggle"
+                    AutomationProperties.Name="Copy GIF to the clipboard"
+                    IsOn="{x:Bind ViewModel.CopyGifToClipboard, Mode=TwoWay}" />
+            </Grid>
+        </Border>
+    </StackPanel>
+</UserControl>
