+++ windows/src/TinyClips.App/Settings/Sections/HotkeysSettingsSection.xaml
@@ -1,0 +1,141 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<UserControl
+    x:Class="TinyClips.App.Settings.Sections.HotkeysSettingsSection"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d">
+
+    <StackPanel Spacing="8">
+        <TextBlock
+            Margin="4,8,0,4"
+            Style="{StaticResource SubtitleTextBlockStyle}"
+            Text="Keyboard shortcuts" />
+
+        <InfoBar
+            IsClosable="False"
+            IsOpen="True"
+            Message="Click Edit, then press the key combination you want (include Ctrl, Alt, Shift, or Win)."
+            Severity="Informational" />
+
+        <Border Style="{StaticResource SettingCardStyle}">
+            <StackPanel Spacing="12">
+                <Grid ColumnSpacing="12">
+                    <Grid.ColumnDefinitions>
+                        <ColumnDefinition Width="*" />
+                        <ColumnDefinition Width="Auto" />
+                        <ColumnDefinition Width="Auto" />
+                        <ColumnDefinition Width="Auto" />
+                    </Grid.ColumnDefinitions>
+                    <TextBlock
+                        VerticalAlignment="Center"
+                        Style="{StaticResource BodyStrongTextBlockStyle}"
+                        Text="Screenshot" />
+                    <Border
+                        Grid.Column="1"
+                        MinWidth="120"
+                        Padding="10,4"
+                        VerticalAlignment="Center"
+                        Background="{ThemeResource ControlFillColorDefaultBrush}"
+                        CornerRadius="{ThemeResource ControlCornerRadius}">
+                        <TextBlock
+                            HorizontalAlignment="Center"
+                            FontFamily="Consolas"
+                            Text="{x:Bind ViewModel.ScreenshotHotKeyDisplay, Mode=OneWay}" />
+                    </Border>
+                    <Button
+                        Grid.Column="2"
+                        AutomationProperties.AutomationId="EditScreenshotHotKeyButton"
+                        AutomationProperties.Name="Edit screenshot shortcut"
+                        Click="OnEditHotKey"
+                        Content="Edit"
+                        Tag="Screenshot" />
+                    <Button
+                        Grid.Column="3"
+                        AutomationProperties.AutomationId="ResetScreenshotHotKeyButton"
+                        AutomationProperties.Name="Reset screenshot shortcut"
+                        Click="OnResetHotKey"
+                        Content="Reset"
+                        Tag="Screenshot" />
+                </Grid>
+                <Grid ColumnSpacing="12">
+                    <Grid.ColumnDefinitions>
+                        <ColumnDefinition Width="*" />
+                        <ColumnDefinition Width="Auto" />
+                        <ColumnDefinition Width="Auto" />
+                        <ColumnDefinition Width="Auto" />
+                    </Grid.ColumnDefinitions>
+                    <TextBlock
+                        VerticalAlignment="Center"
+                        Style="{StaticResource BodyStrongTextBlockStyle}"
+                        Text="Record video" />
+                    <Border
+                        Grid.Column="1"
+                        MinWidth="120"
+                        Padding="10,4"
+                        VerticalAlignment="Center"
+                        Background="{ThemeResource ControlFillColorDefaultBrush}"
+                        CornerRadius="{ThemeResource ControlCornerRadius}">
+                        <TextBlock
+                            HorizontalAlignment="Center"
+                            FontFamily="Consolas"
+                            Text="{x:Bind ViewModel.VideoHotKeyDisplay, Mode=OneWay}" />
+                    </Border>
+                    <Button
+                        Grid.Column="2"
+                        AutomationProperties.AutomationId="EditVideoHotKeyButton"
+                        AutomationProperties.Name="Edit record video shortcut"
+                        Click="OnEditHotKey"
+                        Content="Edit"
+                        Tag="Video" />
+                    <Button
+                        Grid.Column="3"
+                        AutomationProperties.AutomationId="ResetVideoHotKeyButton"
+                        AutomationProperties.Name="Reset record video shortcut"
+                        Click="OnResetHotKey"
+                        Content="Reset"
+                        Tag="Video" />
+                </Grid>
+                <Grid ColumnSpacing="12">
+                    <Grid.ColumnDefinitions>
+                        <ColumnDefinition Width="*" />
+                        <ColumnDefinition Width="Auto" />
+                        <ColumnDefinition Width="Auto" />
+                        <ColumnDefinition Width="Auto" />
+                    </Grid.ColumnDefinitions>
+                    <TextBlock
+                        VerticalAlignment="Center"
+                        Style="{StaticResource BodyStrongTextBlockStyle}"
+                        Text="Record GIF" />
+                    <Border
+                        Grid.Column="1"
+                        MinWidth="120"
+                        Padding="10,4"
+                        VerticalAlignment="Center"
+                        Background="{ThemeResource ControlFillColorDefaultBrush}"
+                        CornerRadius="{ThemeResource ControlCornerRadius}">
+                        <TextBlock
+                            HorizontalAlignment="Center"
+                            FontFamily="Consolas"
+                            Text="{x:Bind ViewModel.GifHotKeyDisplay, Mode=OneWay}" />
+                    </Border>
+                    <Button
+                        Grid.Column="2"
+                        AutomationProperties.AutomationId="EditGifHotKeyButton"
+                        AutomationProperties.Name="Edit record GIF shortcut"
+                        Click="OnEditHotKey"
+                        Content="Edit"
+                        Tag="Gif" />
+                    <Button
+                        Grid.Column="3"
+                        AutomationProperties.AutomationId="ResetGifHotKeyButton"
+                        AutomationProperties.Name="Reset record GIF shortcut"
+                        Click="OnResetHotKey"
+                        Content="Reset"
+                        Tag="Gif" />
+                </Grid>
+            </StackPanel>
+        </Border>
+    </StackPanel>
+</UserControl>
