+++ windows/src/TinyClips.App/Settings/Sections/ScreenshotSettingsSection.xaml
@@ -1,0 +1,153 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<UserControl
+    x:Class="TinyClips.App.Settings.Sections.ScreenshotSettingsSection"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d">
+
+    <StackPanel Spacing="8">
+        <TextBlock
+            Margin="4,8,0,4"
+            Style="{StaticResource SubtitleTextBlockStyle}"
+            Text="Screenshot" />
+
+        <TextBlock
+            Margin="4,4,0,0"
+            Style="{StaticResource BodyStrongTextBlockStyle}"
+            Text="Image &amp; quality" />
+
+        <Border Style="{StaticResource SettingCardStyle}">
+            <Grid ColumnSpacing="16">
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="*" />
+                    <ColumnDefinition Width="Auto" />
+                </Grid.ColumnDefinitions>
+                <StackPanel VerticalAlignment="Center">
+                    <TextBlock Style="{StaticResource BodyStrongTextBlockStyle}" Text="Image format" />
+                </StackPanel>
+                <ComboBox
+                    Grid.Column="1"
+                    VerticalAlignment="Center"
+                    AutomationProperties.AutomationId="ScreenshotFormatComboBox"
+                    AutomationProperties.Name="Screenshot image format"
+                    SelectedIndex="{x:Bind ViewModel.ScreenshotFormatIndex, Mode=TwoWay}">
+                    <ComboBoxItem Content="PNG" />
+                    <ComboBoxItem Content="JPEG" />
+                </ComboBox>
+            </Grid>
+        </Border>
+
+        <Border Style="{StaticResource SettingCardStyle}">
+            <Grid ColumnSpacing="16">
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="*" />
+                    <ColumnDefinition Width="Auto" />
+                </Grid.ColumnDefinitions>
+                <StackPanel VerticalAlignment="Center">
+                    <TextBlock Style="{StaticResource BodyStrongTextBlockStyle}" Text="Scale" />
+                    <TextBlock
+                        Style="{StaticResource CaptionTextBlockStyle}"
+                        Foreground="{ThemeResource TextFillColorSecondaryBrush}"
+                        Text="Resize saved screenshots as a percentage of full resolution." />
+                </StackPanel>
+                <Slider
+                    Grid.Column="1"
+                    Width="220"
+                    VerticalAlignment="Center"
+                    AutomationProperties.AutomationId="ScreenshotScaleSlider"
+                    AutomationProperties.Name="Screenshot scale percentage"
+                    Maximum="100"
+                    Minimum="25"
+                    StepFrequency="5"
+                    TickFrequency="25"
+                    Value="{x:Bind ViewModel.ScreenshotScale, Mode=TwoWay}" />
+            </Grid>
+        </Border>
+
+        <Border Style="{StaticResource SettingCardStyle}">
+            <Grid ColumnSpacing="16">
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="*" />
+                    <ColumnDefinition Width="Auto" />
+                </Grid.ColumnDefinitions>
+                <StackPanel VerticalAlignment="Center">
+                    <TextBlock Style="{StaticResource BodyStrongTextBlockStyle}" Text="JPEG quality" />
+                    <TextBlock
+                        Style="{StaticResource CaptionTextBlockStyle}"
+                        Foreground="{ThemeResource TextFillColorSecondaryBrush}"
+                        Text="Only used when the image format is JPEG." />
+                </StackPanel>
+                <Slider
+                    Grid.Column="1"
+                    Width="220"
+                    VerticalAlignment="Center"
+                    AutomationProperties.AutomationId="JpegQualitySlider"
+                    AutomationProperties.Name="JPEG quality"
+                    Maximum="1.0"
+                    Minimum="0.1"
+                    StepFrequency="0.05"
+                    Value="{x:Bind ViewModel.JpegQuality, Mode=TwoWay}" />
+            </Grid>
+        </Border>
+
+        <TextBlock
+            Margin="4,16,0,0"
+            Style="{StaticResource BodyStrongTextBlockStyle}"
+            Text="Capture behavior" />
+
+        <Border Style="{StaticResource SettingCardStyle}">
+            <Grid ColumnSpacing="16">
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="*" />
+                    <ColumnDefinition Width="Auto" />
+                </Grid.ColumnDefinitions>
+                <StackPanel VerticalAlignment="Center">
+                    <TextBlock Style="{StaticResource BodyStrongTextBlockStyle}" Text="Countdown before capture" />
+                </StackPanel>
+                <StackPanel
+                    Grid.Column="1"
+                    VerticalAlignment="Center"
+                    Orientation="Horizontal"
+                    Spacing="12">
+                    <NumberBox
+                        Width="120"
+                        AutomationProperties.AutomationId="ScreenshotCountdownDurationNumberBox"
+                        AutomationProperties.Name="Screenshot countdown seconds"
+                        IsEnabled="{x:Bind ViewModel.ScreenshotCountdownEnabled, Mode=OneWay}"
+                        Maximum="10"
+                        Minimum="1"
+                        SpinButtonPlacementMode="Inline"
+                        Value="{x:Bind ViewModel.ScreenshotCountdownDuration, Mode=TwoWay}" />
+                    <ToggleSwitch
+                        AutomationProperties.AutomationId="ScreenshotCountdownToggle"
+                        AutomationProperties.Name="Enable screenshot countdown"
+                        IsOn="{x:Bind ViewModel.ScreenshotCountdownEnabled, Mode=TwoWay}" />
+                </StackPanel>
+            </Grid>
+        </Border>
+
+        <Border Style="{StaticResource SettingCardStyle}">
+            <Grid ColumnSpacing="16">
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="*" />
+                    <ColumnDefinition Width="Auto" />
+                </Grid.ColumnDefinitions>
+                <StackPanel VerticalAlignment="Center">
+                    <TextBlock Style="{StaticResource BodyStrongTextBlockStyle}" Text="Open editor after capture" />
+                    <TextBlock
+                        Style="{StaticResource CaptionTextBlockStyle}"
+                        Foreground="{ThemeResource TextFillColorSecondaryBrush}"
+                        Text="Review, crop, and copy each screenshot before saving." />
+                </StackPanel>
+                <ToggleSwitch
+                    Grid.Column="1"
+                    VerticalAlignment="Center"
+                    AutomationProperties.AutomationId="ShowScreenshotEditorToggle"
+                    AutomationProperties.Name="Open editor after capture"
+                    IsOn="{x:Bind ViewModel.ShowScreenshotEditor, Mode=TwoWay}" />
+            </Grid>
+        </Border>
+    </StackPanel>
+</UserControl>
